src/requirements.txt
python 3.6.*

matplotlib
pydensecrf
numpy
Pillow
torch
torchvision
opencv2
pydicom
